requirements.txt
agno==2.3.6
haystack-ai==2.20.0
neo4j_haystack==2.2.1
neo4j==6.0.3
openai==2.8.1
scikit_learn==1.7.2
tqdm==4.67.1
cn2an==0.5.23
loguru==0.7.3
numpy==2.3.5
pandas==2.3.3
sqlalchemy==2.0.44
prompt_toolkit==3.0.52
json_repair==0.54.2